Ahora van en donde corresponden. lo habia puesto donde NO era
